requirements.txt
textblob
pyspellchecker==0.7.2